@@ -2,25 +2,26 @@ name: camera
 description: A Flutter plugin for getting information about and controlling the
   camera on Android and iOS. Supports previewing the camera feed, capturing images, capturing video,
   and streaming image buffers to dart.
-version: 0.7.0+3
+version: 0.8.0-nullsafety
 homepage: https://github.com/flutter/plugins/tree/master/packages/camera/camera
 
 dependencies:
   flutter:
     sdk: flutter
-  camera_platform_interface: ^1.5.0
-  pedantic: ^1.8.0
-  quiver: ^2.1.5
+
+  camera_platform_interface: ^2.0.0-nullsafety
+
+  pedantic: ^1.10.0
+  quiver: ^3.0.0-nullsafety.3
 
 dev_dependencies:
-  path_provider: ^0.5.0
-  video_player: ^0.10.0
+  video_player: ^2.0.0-nullsafety.7
   flutter_test:
     sdk: flutter
   flutter_driver:
     sdk: flutter
-  mockito: ^4.1.3
-  plugin_platform_interface: ^1.0.3
+  mockito: ^5.0.0-nullsafety.5
+  plugin_platform_interface: ^1.1.0-nullsafety.2
 
 flutter:
   plugin:
@@ -32,5 +33,5 @@ flutter:
         pluginClass: CameraPlugin
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.12.13+hotfix.5"
+  sdk: '>=2.12.0-0 <3.0.0'
+  flutter: ">=1.22.0"
